feat: move pipeline parity debug endpoint to a dedicated router for improved organization
