+++ backend-migration/app/framework/api/endpoint_registry.py
@@ -3,6 +3,7 @@
 from fastapi import FastAPI
 
 from app.api.endpoints import metadata
+from app.framework.api.debug_router import create_debug_router
 
 
 def register_default_endpoints(app: FastAPI) -> None:
@@ -13,3 +14,4 @@ def register_default_endpoints(app: FastAPI) -> None:
     the existing endpoint contract by wiring the current metadata router.
     """
     app.include_router(metadata.router)
+    app.include_router(create_debug_router())
